move old network code into a temporary folder
