+++ src/client/game.cpp
@@ -2,8 +2,6 @@
 
 bool Game::initGame()
 {
-    // Client.serverIp = LOCAL_HOST;
-
     m_serverLauncher = std::make_unique<ServerLauncher>(sf::milliseconds(1000));
     m_serverLauncher->runAsThread();
 
@@ -17,28 +15,18 @@ bool Game::initGame(const std::string& ipAddress)
 
 bool Game::init(const std::string& ip)
 {
-    auto result = m_netClient.connectTo(ip);
+    auto result = m_client.connectTo(ip);
     if (!result.success) {
         std::cout << "ERROR: " << result.message << "\n";
-        return false;
-    }
-
-    m_client = std::make_unique<Client>();
-    if (!m_client->init(ip)) {
         stopGame();
         return false;
     }
+    m_isInGame = true;
     return true;
 }
 
 void Game::stopGame()
 {
-    if (m_client) {
-        m_netClient.disconnect();
-        m_client->endGame();
-        m_client->destroy();
-        m_client.release();
-    }
     if (m_serverLauncher) {
         m_serverLauncher->stop();
         m_serverLauncher.release();
@@ -47,30 +35,27 @@ void Game::stopGame()
 
 void Game::onMouseRelease(sf::Mouse::Button button)
 {
-    if (m_client) {
-        m_client->onMouseRelease(button);
+    if (m_isInGame) {
     }
 }
 
 void Game::input(sf::Window& window, const Keyboard& keyboard,
                  const InputState& inputState)
 {
-    if (m_client) {
-        m_client->handleInput(window, keyboard, inputState);
+    if (m_isInGame) {
     }
 }
 
 void Game::update(float dt)
 {
-    if (m_serverLauncher) {
-        m_netClient.tick();
-        // m_client->update(dt);
+    if (m_isInGame) {
+        m_client.tick();
     }
 }
 
 void Game::render()
 {
-    if (m_client) {
-        m_client->render();
+    if (m_isInGame) {
+        
     }
 }
